+++ tests/index.test.ts
@@ -9,13 +9,10 @@ async function test() {
     reconnect: true,
   });
 
-  // const sql = conn.query.select("*").from("citizens").order("full_name", "ASC");
-  const sql = conn.query.raw("SELECT * FROM `citizens`").where("id", "a22e0c79-5ecd-4012-85bb-a214196e2d72");
+  const x = await conn.query().select(["id", "full_name", "user_id"]).from("citizens").order("full_name", "ASC").exec();
+  console.log(x);
 
-  // const sql = conn.query.delete("citizens").where("id", "b33a9b77-3313-43df-a2c2-1e6e20858b81");
-  const result = await sql.exec().catch(console.error);
-
-  result;
+  // const sql = conn.query().delete("citizens").where("id", "b33a9b77-3313-43df-a2c2-1e6e20858b81");
 
   await conn.end();
 }